rename BaseXboxControllerAdapter to XXboxController
